Moved piece test files into appropriate package
