+++ src/test/java/com/stateofflux/chess/model/pieces/PawnMovesTest.java
@@ -1,9 +1,8 @@
-package com.stateofflux.chess.model;
+package com.stateofflux.chess.model.pieces;
 
+import com.stateofflux.chess.model.Board;
+import com.stateofflux.chess.model.Game;
 import org.testng.annotations.Test;
-
-import com.stateofflux.chess.model.pieces.PieceMoves;
-import com.stateofflux.chess.model.pieces.PawnMoves;
 
 import static org.assertj.core.api.Assertions.*;
