+++ src/test/java/com/stateofflux/chess/model/pieces/QueenMovesTest.java
@@ -1,9 +1,7 @@
-package com.stateofflux.chess.model;
+package com.stateofflux.chess.model.pieces;
 
+import com.stateofflux.chess.model.Board;
 import org.testng.annotations.Test;
-
-import com.stateofflux.chess.model.pieces.PieceMoves;
-import com.stateofflux.chess.model.pieces.QueenMoves;
 
 import static org.assertj.core.api.Assertions.*;
