+++ src/test/java/com/stateofflux/chess/model/pieces/RookMovesTest.java
@@ -1,5 +1,6 @@
-package com.stateofflux.chess.model;
+package com.stateofflux.chess.model.pieces;
 
+import com.stateofflux.chess.model.Board;
 import org.testng.annotations.Test;
 
 import com.stateofflux.chess.model.pieces.PieceMoves;
